fixing bug when set triangl param distrub and for bridge alone
If one set min= best estimate < max or the contrary, the distribution computed make the code bugging, I do not know why so we set that min = 0.999 bestestimate or max = 1.001 best estimate in this case.
Also a bug emerged when modelling only bridges because paramters JammingHeight and DepositHeight were undefined. This is fixed.